requirements.txt
streamlit
pandas
gspread
google-auth
plotly

